requirements.txt
pygame
PyOpenGL
pygame
PyOpenGL-accelerate 
numpy 